refactor(weddi.app): move off dialog from board to a unique foler
